refactor: 자동차 도메인 등록
